move TypeCastExpr to types.py
